% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_layer.R
\name{env_layer}
\alias{env_layer}
\title{a climatic envelope layer}
\usage{
env_layer(env, .fill = NA, .colour = "red")
}
\arguments{
\item{env}{a climatic envelope, e.g. \code{env_fasy}}

\item{.fill}{a fill colour}

\item{.colour}{a line colour}
}
\value{
an object of class "gg"
}
\description{
draw a climatic envelope layer into a ggplot
}
