% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env_pcab}
\alias{env_pcab}
\title{Climate envelope for Picea abies}
\format{A \code{data.frame} containing two columns with Prec
  (monthly precipitation sum), and Temp (monthly temperature
  mean).}
\source{
Kölling C. 2007. Klimahüllen von 27 Waldbaumarten. AFZ-Der
  Wald 62: 1242–1245.
}
\usage{
data(env_pcab)
}
\description{
This dataset is the climatic envelope of Picea abies after the
  "Bohn-Map", redrawn from Kölling et al.
}
\keyword{datasets}
